change problem set list permissions
code review changes
